% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOtop_CreateInptsMAPS.R
\name{GEOtop_CreateInptsMAPS}
\alias{GEOtop_CreateInptsMAPS}
\title{Create GEOtop input maps}
\usage{
GEOtop_CreateInptsMAPS(dem, res, stream_gauge, name_gauge, preprocess = TRUE,
  soil = FALSE, landcover = FALSE, bedrock = FALSE, SAGAENV,
  mask_only_lc = FALSE)
}
\arguments{
\item{dem}{ascii digital elevation model, located in working directory}

\item{res}{resolution of the output maps in meters}

\item{stream_gauge}{ascii, defining location of stream gauge, located in working directory (NA ascii, stream gauge location = -1)}

\item{name_gauge}{name of the stream gauge}

\item{preprocess}{boolean, TRUE: preprocessing (terrain analysis); FALSE: no preprocessing, use maps created in former run; default = TRUE}

\item{soil}{boolean, TRUE: soil type map provided in working directory, ascii format, name: soil.asc; FALSE: no soil type map provided; default = FALSE}

\item{landcover}{boolean, TRUE: landcover map provided in working directory, ascii format, name: landcover.asc; FALSE: no landcover map provided; default = FALSE}

\item{bedrock}{boolean, TRUE: bedrock map provided in working directory, ascii format, name: soil.asc; FALSE: no bedrock map provided; default = FALSE}

\item{SAGAENV}{environment, created with function \code{\link{rsaga.env}} of the \pkg{RSAGA} package}

\item{mask_only_lc}{boolean, TRUE: mask only landcover map; FALSE: mask all maps and crop to data; default = FALSE}
}
\value{
GEOtop input maps, storged in working directory as ascii maps

Johannes Brenner, \email{Johannes.Brenner@eurac.edu}
}
\description{
Create GEOtop input maps using \pkg{RSAGA}. Base maps are a digital elevation model and a map locating the basin outlet (defined by -1). Postprocessing includes terrain analysis (slope, aspect, curvature) and sky view factor calculation. All maps, including landcover, soil type and bedrock (if provided), are stored in SAGA .srgd format. Further processing includes watershed delineation, resampling to required resolution, masking to the extent of the watershed (either for all maps or for landcover map only) and writing of GEOtop input ascii files.
}
\details{
GEOtop_CreateInptsMAPS
}

