% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOtop_ReadMultiPoint.R
\name{GEOtop_Read_multipoint}
\alias{GEOtop_Read_multipoint}
\title{GEOtop_Read_multiplot}
\usage{
GEOtop_Read_multipoint(path, model_run, stations, val_aggr, soil_files,
  lc_classes, linux)
}
\arguments{
\item{path}{path pointing to simulation folder}

\item{model_run}{name of simulation folder}

\item{stations}{character vector, names of GEOtop output sites}

\item{val_aggr}{aggregation performed on output data. "h": hourly; "d": daily}

\item{soil_files}{boolean, TRUE: soil files are provided as GEOtop input. FALSE: soil is parameterized in the geotop.inpts file}

\item{lc_classes}{character vector, names of landcover classes in GEOtop simulation}

\item{linux}{boolean, TRUE: working on linux laptop, FALSE: working on windows laptop; for path retrieval via \emph{data(SWCinfo)}}
}
\description{
Read multi point output from GEOtop simulation using the \pkg{geotopbricks} package.
}
\details{
Read multi point output from GEOtop simulation
}
\note{
Slow, as point files are read for each variable!
}
\author{
Johannes Brenner, \email{Johannes.Brenner@eurac.edu}
}
\references{
Endrizzi, S., Gruber, S., Amico, M. D., & Rigon, R. (2014). \strong{GEOtop 2.0 : simulating the combined energy and water balance at and below the land surface accounting for soil freezing , snow cover and terrain effects.} \emph{Geosci. Model Dev., 7, 2831-2857}. doi:10.5194/gmd-7-2831-2014
}

