% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOtop_ReadPlotRst.R
\name{GEOtop_ReadPlotRst}
\alias{GEOtop_ReadPlotRst}
\title{Visualize asc raster map}
\usage{
GEOtop_ReadPlotRst(map, date, variable, layer, limits, legend,
  lowcol = "#f7fbff", highcol = "#08306b")
}
\arguments{
\item{map}{character, full path and name of .asc map to visualize}

\item{date}{character, date characterization}

\item{variable}{character, variable characterization}

\item{layer}{layer to plot}

\item{limits}{limits to plot}

\item{legend}{character, legend characterization, e.g. \code{"vol\%"} for soil moisture data}

\item{lowcol}{minimum colour for visualization, default is \code{"#f7fbff"}}

\item{highcol}{maximum colour for visualization, default is \code{"#08306b"}; default is creating a blue colour range}
}
\value{
ggplot object
}
\description{
Visualize .asc raster map (GEOtop map output) with ggplot
}
\author{
Johannes Brenner \email{Johannes.Brenner@eurac.edu}
}

