% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOtop_ReadPointData.R
\name{GEOtop_ReadPointData}
\alias{GEOtop_ReadPointData}
\title{Function to load GEOtop point simulation output based on observations}
\usage{
GEOtop_ReadPointData(wpath, soil_output_files = c("SoilLiqContentProfileFile",
  "SoilIceContentProfileFile", "SoilLiqWaterPressProfileFile",
  "SoilAveragedTempProfileFile"), soil_files = TRUE, save_rData = TRUE)
}
\arguments{
\item{wpath}{path pointing into simulation folder}

\item{soil_output_files}{output file keywords used in the function. Default is \code{c("SoilLiqContentProfileFile","SoilIceContentProfileFile", "SoilLiqWaterPressProfileFile", "SoilAveragedTempProfileFile")}}

\item{soil_files}{boolean, \code{TRUE}: soil files are provided as GEOtop input. \code{FALSE}: soil is parameterized in the geotop.inpts file}

\item{save_rData}{boolean, if \code{TRUE} (default) data is stored in working directory (simulation folder)}
}
\description{
Function to load GEOtop point simulation output based on observations
}
\examples{
 
wpath <- system.file('simulation/Muntatschini_pnt_1_225_B2_004',package="AnalyseGeotop")
data("observations_B2")

out <- GEOtop_ReadPointData(wpath, soil_files=TRUE, save_rData=TRUE)


}

