% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOtop_ReadPointVar.R
\name{GEOtop_ReadPointVar}
\alias{GEOtop_ReadPointVar}
\title{Read specific variable from GEOtop point output
	
Read specific variable from GEOtop point output, e.g. precipitation from PointOutputFile, SMC of specific layer PointOutputFile.}
\usage{
GEOtop_ReadPointVar(wpath, keyword, varOFint, coordinates_value_df = FALSE,
  landcover_classes)
}
\arguments{
\item{wpath}{path pointing into simulation folder}

\item{keyword}{keyword for files to read (point: "PointOutputFile", snow: "SnowProfileFile", snowcover: "SnowCoveredAreaFile", thetaliq: "SoilLiqContentProfileFile", thetaliq: "SoilIceContentProfileFile")}

\item{varOFint}{keyword for variable of interest (see headers of GEOtop output files)}

\item{coordinates_value_df}{boolean, TRUE: create data frame containing output coordinates and corresponding characteristics (input maps have to be provided, default = FALSE)}

\item{landcover_classes}{character vector, names of landcover classes in GEOtop simulation}
}
\description{
Read specific variable from GEOtop point output
	
Read specific variable from GEOtop point output, e.g. precipitation from PointOutputFile, SMC of specific layer PointOutputFile.
}
\author{
Johannes Brenner, \email{Johannes.Brenner@eurac.edu}
}

