% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOtop_VisSoilWaterRet.R
\name{GEOtop_VisSoilWaterRet}
\alias{GEOtop_VisSoilWaterRet}
\title{Visualisation Soil Water Retention Curve
	
Visulisation of the Soil Water Retention Curve and hydraulic conductivity for model used in GEOtop (van Genuchten 1980, Mualem ...). Theoretical and parametererized physical soil water parameters are compared. For soil water and hydraulic conductivity calculation from soil water pressure head (hPa or cm water column) the package \pkg{soilwater} is used.}
\usage{
GEOtop_VisSoilWaterRet(alpha, n, theta_sat, theta_res, theta_pwp, theta_fc,
  observed = NULL, add_ref_curves = TRUE, pdf = FALSE, png = TRUE,
  accurate = 10, ksat, colors = NULL, plot_khy = TRUE, do.legend = TRUE)
}
\arguments{
\item{alpha}{alpha van Genuchten parameter [1/m]; inverse of a length - scale parameter}

\item{n}{N van Genuchten parameter [-]; shape parameter}

\item{theta_sat}{saturated soil water content [m³/m³]}

\item{theta_res}{residual soil water content [m³/m³]}

\item{theta_pwp}{permanent wilting point [m³/m³]}

\item{observed}{numeric matrix, providing observed soil moisture data [m³/m³] (first column) and water potential [hPa] (second column)}

\item{add_ref_curves}{boolean, default = TRUE: reference soil water retention curves are added for clay, loam and sand; van Genuchten parameterisation according to ROSETTA manual}

\item{pdf}{boolean, default = FALSE: visualisation is saved in pdf if TRUE}

\item{png}{boolean, default = TRUE: visualisation is saved in png if TRUE; if png=FALSE and pdf=FALSE vis is ploted in x11()}

\item{accurate}{default = 10: water potential is given as \emph{seq(1,10^7,10)}, for accurate plotting set argument 1}

\item{ksat}{saturated hydraulic conductivity [mm/s]}

\item{colors}{color spezification for water retention curves, default = NULL: grey color range is used}

\item{plot_khy}{boolean, default = TRUE: plot Mualem model - relation between hydraulic conducivity and water potential}

\item{thata_fc}{field capacity [m³/m³]}
}
\description{
Visualisation Soil Water Retention Curve
	
Visulisation of the Soil Water Retention Curve and hydraulic conductivity for model used in GEOtop (van Genuchten 1980, Mualem ...). Theoretical and parametererized physical soil water parameters are compared. For soil water and hydraulic conductivity calculation from soil water pressure head (hPa or cm water column) the package \pkg{soilwater} is used.
}
\details{
If multiple vanGenuchten - Mualem models sould be visualised, simply provide alpha, n, theta_sat, theta_res, and ksat values as vectors. Provide alpha as a named vector in order to specify legend names. 


 visualize soil water retention curve and GEOtop model parameterization
 using swc function from soilwater

 data from point.RData (GEOtop_ReadMultiPoint)

 wpath="/run/user/1000/gvfs/smb-shnare:server=sdcalp01.eurac.edu,share=data2/Simulations/Simulation_GEOtop_1_225_ZH/Vinschgau/SimTraining/BrJ/HiResAlp/1D/Montecini_pnt_1_225_B2_003"

wpath="/run/user/1000/gvfs/smb-share:server=sdcalp01.eurac.edu,share=data2/Simulations/Simulation_GEOtop_1_225_ZH/Vinschgau/SimTraining/BrJ/HiResAlp/1D/Montecini_pnt_1_225_P2_002/"

library(geotopbricks)
library(soilwater)

soil_input <- get.geotop.inpts.keyword.value(keyword="SoilParFile", wpath=wpath, data.frame=TRUE)

alpha <- soil_input$alpha[1]
n <- soil_input$n[1]
theta_sat <- soil_input$vwc_s[1]
theta_res <- soil_input$vwc_r[1]
theta_pwp <- soil_input$vwc_w[1]
theta_fc <- soil_input$vwc_fc[1]
ksat <- soil_input$Kh[1]
}

