% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_Caldero.R
\docType{data}
\name{observation_Caldero}
\alias{observation_Caldero}
\title{observation_Caldero}
\format{data.frame}
\source{
www.eurac.edu
}
\usage{
data(observation_Caldero)
}
\description{
lookup_tbl_observation ...WRITE A DESCRIPTION
}
\author{
Johannes Brenner
}

