% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_Caldero.R
\docType{data}
\name{observations_Caldero}
\alias{Caldero_30min}
\alias{Caldero_d}
\alias{observations_Caldero}
\title{observations_Caldero}
\format{list containing zoo objects (half hourly and daily data)}
\source{
www.eurac.edu
}
\usage{
data(observations_Caldero)
}
\description{
lookup_tbl_observation Observational data Caldaro/Kaltern station, half hourly and daily data aggregation
}
\author{
Johannes Brenner
}

