% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed_B2_d.R
\docType{data}
\name{observed_B2_d}
\alias{observed_B2_d}
\title{observed_B2_d}
\format{data.frame}
\source{
www.eurac.edu
}
\usage{
data(observed_B2_d)
}
\description{
lookup_tbl_observation ...WRITE A DESCRIPTION
}
\author{
Johannes Brenner
}

