% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed_B2_h.R, R/validation.R
\docType{data}
\name{observed_B2_h}
\alias{observed_B2_h}
\title{observed_B2_h}
\format{data.frame}
\source{
www.eurac.edu

www.eurac.edu
}
\usage{
data(observed_B2_h)

data(observed_B2_h)
}
\description{
...WRITE A DESCRIPTION

...WRITE A DESCRIPTION
}
\author{
Johannes Brenner

Johannes Brenner
}

