% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWCInfo.R
\docType{data}
\name{SWCinfo}
\alias{SWCinfo}
\alias{SWCinfoLIN}
\alias{SWCinfoWIN}
\title{SWCinfo}
\format{data.frame}
\source{
www.eurac.edu
}
\usage{
data(SWCinfo)
}
\description{
...WRITE A DESCRIPTION
}
\author{
Johannes Brenner
}

