% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOtop_ReadPointData.R
\name{GEOtop_ReadPointData}
\alias{GEOtop_ReadPointData}
\title{Function to load GEOtop point simulation output based on observations}
\usage{
GEOtop_ReadPointData(wpath, soil_output_files = c("SoilLiqContentProfileFile",
  "SoilIceContentProfileFile", "SoilLiqWaterPressProfileFile",
  "SoilAveragedTempProfileFile"), soil_files = TRUE, save_rData = TRUE)
}
\description{
Function to load GEOtop point simulation output based on observations
}
\examples{
 
wpath <- "/run/user/1000/gvfs/smb-share:server=sdcalp01.eurac.edu,
share=data2/Simulations/Simulation_GEOtop_1_225_ZH/Vinschgau/SimTraining/BrJ/
HiResAlp/1D/Montecini_pnt_1_225_B2_007/"
wpath <- "/run/user/1000/gvfs/smb-share:server=sdcalp01.eurac.edu,
share=data2/Simulations/Simulation_GEOtop_1_225_ZH/Vinschgau/SimTraining/BrJ/
MonaLisa/1D/Kaltern/sim006"
data("observations_B2")

load(file.path(wpath, "obs", "observation.RData"))
names(observation) <- c("hour", "day")
obs <- observation

obs   <- list(hour=B2_h, day=B2_d)


}

