% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_tbl_observation.R
\docType{data}
\name{lookup_tbl_observation}
\alias{lookup_tbl_observation}
\title{lookup_tbl_observation}
\format{data.frame}
\source{
www.eurac.edu
}
\usage{
data(lookup_tbl_observation)
}
\description{
lookup_tbl_observation ...WRITE A DESCRIPTION
}
\author{
Johannes Brenner
}

