% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_B2.R
\docType{data}
\name{observation_B2}
\alias{observation_B2}
\title{observation_B2}
\format{data.frame}
\source{
www.eurac.edu
}
\usage{
data(observation_B2)
}
\description{
lookup_tbl_observation ...WRITE A DESCRIPTION
}
\author{
Johannes Brenner
}

